requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.9
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pycodestyle==2.10.0
pymongo==3.11.0
pytz==2023.3
sqlparse==0.2.4
tzdata==2023.3
